- added unittest for sharding module - fixed shard test/example datatype to "Float" since Numeric behavior has changed - added docstrings to sqlalchemy/orm/__init__.py functions; still need to proof within generated HTML pages/fix errors - added engine.threadlocal to docs, removed old mods.threadlocal

+++ examples/sharding/attribute_shard.py
@@ -77,7 +77,7 @@ weather_locations = Table("weather_locations", meta,
 weather_reports = Table("weather_reports", meta,
     Column('id', Integer, primary_key=True),
     Column('location_id', Integer, ForeignKey('weather_locations.id')),
-    Column('temperature', Numeric),
+    Column('temperature', Float),
     Column('report_time', DateTime, default=datetime.datetime.now),
 )
